refactoring started
