+++ package.json
@@ -17,6 +17,7 @@
     "react": "18.3.1",
     "react-dom": "18.3.1",
     "react-router-dom": "^6.27.0",
+    "react-tooltip": "^5.28.0",
     "recoil": "^0.7.7",
     "uuid": "^11.0.2"
   },
